Remove remaining legacy scripts files
